==> Genomgångar/01_SeriesDataframes.ipynb <==
# Series
- är en endimentionell array av vilken datatyp som helst

import pandas as pd

fruits = pd.Series(["kiwi", "kumqwat", "banan", "melon", "mango", "apple", "strawberry"], index=["a", "b", "c", "d", "e", "f", "g"])
numbers = pd.Series([1, 2, 3, 4, 5, 6, 7, 8, 9])

numbers[numbers > 5]

# Dataframe
- är en tvådimentionell datastruktur

- **Framförallt detta vi kommer jobba med!**

data = {
    "Name": ["John", "Anna", "Peter", "Linda"],
    "Location": ["New York", "Paris", "Berlin", "London"],
    "Age": [24, 13, 53, 33],
}

df = pd.DataFrame(data)
df


age = df["Age"]
datingregler = age / 2 + 7
datingregler

df[df["Age"] > 30]

str(df)
