==> Genomgångar/03_import_data.ipynb <==
# Dataframes från filer

- pandas (pd) kan hantera data från många filtyper. typ excel! förutom de vanliga också då såklart (csv, json osv)
- excel behöver biblioteket openpyxl

- se längre ner för smidiga sätt att hitta missad data osv

import pandas as pd
import seaborn as sns

pokemons = pd.read_csv("../data/datapoints.csv")

pokemons.head(10)

salaries = pd.read_csv("../data/salaries.csv")
salaries

salaries.head(10)
salaries.tail(10)
salaries.sample(10)
pd.set_option('display.max_rows', 20)
rows, columns = salaries.shape
print(f"Salaries has {rows} rows and {columns} columns")

salaries.info()

# sns.relplot(data=salaries, x="salary_in_usd", y="company_size", hue="job_title")

### Kopiera dataframes till ctrl+c



salaries.to_clipboard()

### Ändra filtyp

- Finns olika sätt att spara. Kolla doc och Fredriks genomgång

sal_json = salaries.to_json(orient="records")
# salaries.to_json("../data/salaries.json")

# salaries.to_dict(orient="records")
