==> Genomgångar/04_Loc_and_iloc.ipynb <==
# Loc, iloc och []

- Pandas hanterar tre sorters multipel-axels-indexering för object väljning.
    - **loc** används för labelbased indexering
    - **iloc** används för integer based indexering
    - **[kolumnnamn]** returnerar kolumnen som en pandas Series objekt

import pandas as pd

column_values = [f"Column {i}" for i in range(1, 5)]
index_values = [chr(i) for i in range(65, 72)] # chr() hämtar tecken från ASCII-tabellen

df = pd.DataFrame([[f"{ind}{col}" for col in range(1, 5)] for ind in index_values], columns=column_values, index=index_values)
df

df.iloc[1, 0] # Detta ger oss värdet i rad 1, kolumn 0. Funkar även med negativ indexering.
# iloc kan också använda slices ANNORLUNDA I LOC! (inclusive/exclusive)

df.loc["B", "Column 1"] # Detta ger oss värdet i rad B, kolumn "Column 1".

df.iloc[:, 1] # Detta ger oss alla värden i kolumn 1.
df.iloc[1, :] # Detta ger oss alla värden i rad 1.

df.loc["B", :] # Detta ger oss alla värden i rad B.
df.loc[:, "Column 1"] # Detta ger oss alla värden i kolumn "Column 1".

df.loc[:"C", "Column 2"]

# att slicea båda axlarna returnerar en Dataframe
df.loc[:"C", "Column 2":]

df.loc["C":"F", "Column 2":"Column 3"]

df.iloc[2:6, 1:3]

df.iloc[:]

df.loc[:"B", :"Column 2"]

# istället för slices, anävnd listor:

df.loc[["A", "C", "E"], ["Column 1", "Column 3"]]

# **[colnamn]**
Syntax:
- pandasvariabel[colname]
- returnerar en Series object i pandas

df["Column 2"]

df[["Column 2", "Column 4", "Column 1", "Column 1", "Column 1"]]

df["Column 2"]["C"]
