==> Genomgångar/05_Update_Add_Delete.ipynb <==
# Updatera celler med loc och iloc

- En eller flera celler kan sättas till nya värden med loc och iloc

import pandas as pd

column_values = [f"Column {i}" for i in range(1, 5)]
index_values = [chr(i) for i in range(65, 72)] # chr() hämtar tecken från ASCII-tabellen

df = pd.DataFrame([[f"{ind}{col}" for col in range(1, 5)] for ind in index_values], columns=column_values, index=index_values)
df

copy = df.copy()

copy.iloc[[1, 3, 5], [0, 2]] = "NY"
copy

copy.iloc[:2, :2] = copy.iloc[-2:, -2:] + "extra"
copy

## lägg till rader eller kolumner

- Lägg till med endast loc **men inte iloc**

copy = df.copy()

copy.loc["H", :] = ["H1", None, "H3", "H4"] # Måste matcha antalet columner

copy

copy.loc[:, "Column 5"] = 0
copy

copy["Column 6"] = copy["Column 3"] + "-" + copy["Column 4"]
copy

### Döp om kolumner och indexes

För dessa måste du skriva ut varje nyt namn och matcha hur många det var:
- .columns
- .index

Använd annars **rename**

copy = df.copy()

list(copy.columns)
copy.columns = ["A", "B", "C", "D"]

copy

copy.index = [0, 1, 2, 3, 4, 5, 6]

copy

# DETTA SPARAS INTE! Måste sparas i en variabel.
# Även argumentet inplace=True istället för att spara en ny variabel. Tar mindre minne.

copy.rename(columns={"A": "Column 1", "B": "Column 2"}, index={0: "A", 1: "B", 2: "C", 3: "D", 4: "E", 5: "F", 6: "G"})


copy.rename(index={"A": "A1", "B": "B1", "C": "C1", "D": "D1", "E": "E1", "F": "F1", "G": "G1"})

# Man kan skriva columns= och index= i sammma rad.

### Ta bort kolumner och rader

copy = df.copy()

copy.drop(index=["A", "C", "E"], columns=["Column 1", "Column 3"], inplace=True)
copy
