==> Genomgångar/09_Encoding_Cat_Features.ipynb <==
# Encoding categorical features

Categorical features är icke-numerska features med limiterat antal val. (oof svengelska)

Innan vi kan föra in dem i de flesta ML algoritmer måste vi göra dem numeriska antingen genom:
- **Label encoding**: Varje unik kategori blir ett nummer, så som 0,1,2,3,4 osv (som att pichu var 0 och pikachu var 1)
- **One-hot encoding**: Varje kategori får ett binary feature. Alltså:
    - Där vi i titanic exemplet hade tre olika reseklasser, istället för at kalla dem 0,1,2 som label enoding hade gjort så delar vi in dem i tre nya kolumner där 0 eller 1 visas beroende på om objektet i fråga var en del av den klassen.

import pandas as pd
import seaborn as sns

titanic = sns.load_dataset("titanic")
titanic.head(1)

titanic.info()

# Skapa en lista med kolumnerna som är datatyperna object och category

cat_columns = [col for col in titanic.columns if titanic[col].dtype in ['object', 'category']]
cat_columns

# Skapa en dataframe med de kategorierna

df_categories = titanic[cat_columns]
df_categories.head(5)

for cat in cat_columns:
    print(f"{cat}: {df_categories[cat].unique()}")

## Label enconding

Använd label encoding om:

- Kategorier har en naturlig ordning (typ minst till störst)
- Det bara finns två kategorier (typ man och kvinna)
- om One-Hot skulle ge för många kolumner

# manual mapping
titanic["embarked"] = titanic["embarked"].map({"S": 0, "C": 1, "Q": 2})

### Auto-mapping med Label encoder från Scikit-learn

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()

titanic["deck"] = le.fit_transform(titanic["deck"])

titanic.head(5)

## One-Hot encoding

Använd one-hot OM:
- Kategorier INTE har en naturlig ordning. (typ länder)
- Antal av kategorier är liten.

titanic.to_json("../Data/titanic_encoded.json", orient="records")

#titanic = pd.get_dummies(data=titanic, columns=["embark_town"])

#------------------------------------------------------------------------
# ERRORN BLIR NÄR DU FÖRSÖKER GÖRA DETTA PÅ SAMMA DataFrame FLERA GÅNGER!
#------------------------------------------------------------------------

titanic.head(5)

cat_columns = [col for col in titanic.columns if titanic[col].dtype in ['object', 'category', 'bool']]
cat_columns

for column in cat_columns:
    titanic[column] = le.fit_transform(titanic[column])

titanic.head(5)
