==> Genomgångar/10_Correlation_analysis.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np

titanic = pd.read_json('../data/titanic_encoded.json')
titanic.head()

# Korrelationsanalys

Detta är en statistisk teknik som användsför att styrkor, likheter och riktiningar som sker mellan två eller fler variabler.

### Pandas .corr() metod

Ger en ny df med korreleationskoefficiaten (1 = max) där alla kolumner hamnar även som rader och jämför värdet mellan alla kolumner. (Oj, mustig mening)

- .abs() = ändrar till det absoluta värdet (det blir + om det var minus)

correlation_matrix = titanic.corr().abs()
correlation_matrix

# Så allt som är över 0.85 är starkt korrelerat
threshold = 0.85

upper = correlation_matrix.where(np.triu(np.ones(correlation_matrix.shape), k=1).astype(bool))

upper.style.map(lambda x: "font-weight: bold" if x > threshold else "")

# Krävde jinja2

columns_to_drop = []

for i in range(len(correlation_matrix.columns)):
    column_name = correlation_matrix.columns[i]
    
    for j in range(i):
        if correlation_matrix.iloc[i, j] > threshold:
            columns_to_drop.append(column_name)
            break

columns_to_drop

titanic.drop(columns=columns_to_drop)
