==> Genomgångar/12_Read_html.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns

boxoffice = pd.read_html("https://en.wikipedia.org/wiki/List_of_Pixar_films", header=1, match="Box office gross")[0]


boxoffice.drop(columns="Ref.", inplace=True)

boxoffice.replace(to_replace="[$]", value="", regex=True, inplace=True)

boxoffice.replace(to_replace=",", value="", regex=True, inplace=True)

boxoffice2 = boxoffice.copy()
boxoffice2

boxoffice2.replace(to_replace="\xa0million", value="000000", regex=True, inplace=True)

boxoffice2

boxoffice2.drop(columns="U.S. and Canada", inplace=True)

boxoffice2.drop(columns="Other territories", inplace=True)

boxoffice2

boxoffice2.replace(to_replace="—", value=np.NaN, inplace=True)

boxoffice2["Worldwide"].astype("int64")
boxoffice2.info()

boxoffice2["Worldwide"] = boxoffice2["Worldwide"].astype("int64")
boxoffice2.info()

boxoffice2["Budget"] = boxoffice2["Budget"].astype("int64") if 
boxoffice2.info()
