==> Genomgångar/14_plotlyEx_layout_styling.ipynb <==
import plotly.express as px

stocks = px.data.stocks()
stocks.head()

px.line(stocks, x='date', y=['GOOG', "AMZN", "MSFT"])

fig = px.line(stocks.rename(columns=dict(GOOG="Google", AMZN="Amazon", MSFT="Microsoft")),
            x='date', y=['Google', "Amazon", "Microsoft"],
            template='plotly_dark', # 'ggplot2', 'seaborn', 'simple_white', 'plotly', 'plotly_white', 'plotly_dark', 'presentation', 'xgridoff', 'ygridoff', 'gridon', 'none'
            labels=dict(variable="Company", date="Date", value="Value"),
            hover_name="variable", hover_data=dict(variable=False),
        )

fig.update_layout(font_family="Courier New", font_size=14, title_font_family="Times New Roman", title_font_size=24,
                    legend_title_font_size=18, legend_font_size=16, legend=dict(orientation="h", title="", x=0.05, y=1.1),
                )

fig.show()
