==> Genomgångar/15_CombiningDatasets.ipynb <==
# Kombinera dataset

Pandas har flera sätt att kombinera dataset

- Concat: kombinera DataFrames
- Join: kombinera på en rad eller kolumn
- Merge: Kombinera liknande? rader och kolumner

import pandas as pd
from helpers import sample_df, hdisplay, nowrap_display



## Concatenate

left = sample_df("A0", "D3", prefix="L_")
right = sample_df("A0", "D3", prefix="R_")

hdisplay([left, right], ["left", "right"])

pd.concat([left, right], axis="columns") # default axis=index

pd.concat([left, right]).reset_index(drop=True)
pd.concat([left, right]).set_index("C")
pd.concat([left, right], axis="columns", keys=["left", "right"])
pd.concat([left, right], ignore_index=True)

left = sample_df("A0", "D3", prefix="L_")
right = sample_df("C2", "F5", prefix="R_")

hdisplay([left, right], ["left", "right"])

pd.concat([left, right], axis="columns")

df = pd.concat([left, right], keys=["left", "right"])
df.loc['left']

## Join
- Tänk på att Vänster/Höger är en grej i syntax. (inte bara nu i våra tillfälliga variabelnamn)

left = sample_df("A0", "D3", prefix="L_").add_suffix("_left")
right = sample_df("C2", "F5", prefix="R_").add_suffix("_right")

hdisplay([left, right], ["left", "right"])

left.join(right, how="right") # finns även inner(tar bara med non-NaN r/k), outer(tar med allt), cross(tar med alla kombos len(r)*len(l)) (left default)

## Merge
- Lite som join, men mer valbart om hur dataframsen ska sättas ihop

- how= finns här också men här är istället default "inner"

- Lätt att blandas ihop med SQL syntax där SQLs merge heter .join

left = sample_df("A0", "D3")
left.loc[:, "F"] = ["F8", "F9", "F10", "F11"]
right = sample_df("F10", "J13")
hdisplay([left, right], ["left", "right"])

left.merge(right, how="outer")

## Lite riktiga exempel

employees = pd.read_json("../Data/employees.json")
departments = pd.read_json("../Data/departments.json")
display(employees)
display(departments)

df = employees.merge(departments, left_on="department", right_on="department_name", how="left", sort=True).drop(columns="department_name")
nowrap_display(df)

users = pd.read_json("../Data/users.json")
users

df.insert(9, "username", 
        df["first_name"].apply(lambda name: name[:3].lower()) + 
        df["last_name"].apply(lambda name: name[:3].lower())
    )

df

nowrap_display(df.merge(users, on="username", how="left"))

new_df = df.copy()

new_df.groupby(["department", "department_head", "location", "office_number", "budget"])["salary"].mean()
