==> Genomgångar/16_Python_DateTime.ipynb <==
# Datetime i python

from datetime import datetime, timedelta, date, time

now = datetime.now()
now

now.second

now.strftime('%Y-%m-%d %H:%M:%S')

now.strptime('2019-01-01 12:00:00', '%Y-%m-%d %H:%M:%S')

### Timedelta

next_year = now + timedelta(days=365)
next_year
