==> Genomgångar/20_Anonymisering.ipynb <==
# Anonymisering

- **pseudonymisering**: Ersätt information som ska döljas med info som gör att urspringsdatan går att återfå genom att göra steget baklänges.
- **anonymisering**: Ersätt datat på ett sätt med intentionen att inte kunna återfå informationen.

### Tekniker

- **Data masking** : "hej på dig" blir "h** p* d**"
- **Hashing / crypto**: "hej på dig" blir /&"#¤IB)#/¤!"#HD)
- **Generalisation**: "hej på dig" blir "hälsning preposition någon"
- **Removing data**: "hej på dig" blir ""
- **Aggregering**: "hej på dig" blir "heeeeeejjjjjjj pååååååååå diiiiiiiiiiiig"

# HASH

import hashlib as hl
import pandas as pd


text = "superhemliga saker".encode()
print(text)

crypted = hl.sha256(text).hexdigest()
crypted

# decrypted = hl.sha256(text).hexdigest()
# decrypted

df = pd.read_json("../Data/salaries.json")
df

df["employee_residence"] = df["employee_residence"].apply(lambda x: hl.sha256(x.encode()).hexdigest())

df.head()

location = input("Plaese enter your location: ")
location_hash = hl.sha256(location.encode()).hexdigest()
print(location_hash)
df.query("employee_residence == @location_hash")
